Refactor to move fetchHandler to utility file
